+++ include/render/WavefrontIntersectionBackend.h
@@ -31,6 +31,7 @@ namespace render {
     std::uint64_t planes{0};
     std::uint64_t rectangles{0};
     std::uint64_t disks{0};
+    std::uint64_t openCylinders{0};
     std::uint64_t transforms{0};
     std::uint64_t unsupportedPrimitives{0};
     std::uint64_t uploadBytes{0};